+++ activesupport/lib/active_support/core_ext/uri.rb
@@ -20,7 +20,7 @@ module URI
   class << self
     def parser
       ActiveSupport::Deprecation.warn(<<-MSG.squish)
-        URI.parser is deprecated and will be removed in Rails 6.2.
+        URI.parser is deprecated and will be removed in Rails 7.0.
         Use `URI::DEFAULT_PARSER` instead.
       MSG
       URI::DEFAULT_PARSER